Make entity IDs nullable and update key generation
Changed Id properties in Contractor, Document, and DocumentItem models and DTOs to nullable int to support auto-generated keys. Updated EF Core model configuration and SQL schema to use SERIAL/auto-increment for primary keys. Adjusted usages and tests to handle nullable IDs and explicit casting where needed. Also fixed project reference path and added Migrations folder to the project file.

+++ WarehouseApi.Tests/DocumentItemsControllerTests.cs
@@ -124,7 +124,7 @@ namespace WarehouseApplication.Tests.Controllers
             var controller = new DocumentItemsController(_contextMock.Object, _mapper);
 
             // Act
-            var result = await controller.Put(existing.Id, dto);
+            var result = await controller.Put((int)existing.Id, dto);
 
             // Assert
             Assert.IsType<NoContentResult>(result);
